requirements.txt
appnope==0.1.2
asgiref==3.3.1
asn1crypto==1.4.0
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cssselect==1.1.0
decorator==4.4.2
Django==3.1.7
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-redis==4.12.1
django-redis-sessions==0.6.2
djangorestframework==3.12.2
djangorestframework-sav==0.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.7.0
incremental==17.5.0
ipython==7.20.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
olefile==0.46
parsel==1.6.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
Protego==0.1.16
psycopg2==2.7.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.8.0
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pytz==2021.1
queuelib==1.5.0
redis==3.5.3
requests==2.25.1
Scrapy==2.4.1
service-identity==18.1.0
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
Twisted==20.3.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
w3lib==1.22.0
wcwidth==0.2.5
zipp==3.4.0
zope.interface==5.2.0
